Continous target variable prediction showing Project
Edureka Certification Project - Social Domain. Mashable (www.mashable.com) -- is a global, multi-platform media and entertainment
company. Powered by its own proprietary technology, Mashable is the go-to source for tech,
digital culture and entertainment content for its dedicated and influential audience around
the globe.
Just like any other media company its success depends on the popularity of articles. And one
of the key metrics to measure popularity is no. of shares done on article. I as a ML expert have to do analysis and modeling to predict number of shares of an article
given the input parameters.